+++ app/src/main/java/com/example/cardinalgate/core/helper/ByteTool.kt
@@ -1,0 +1,7 @@
+package com.example.cardinalgate.core.helper
+
+class ByteTool {
+    companion object {
+        fun b(byte: Int) = byte.toByte()
+    }
+}