+++ app/src/main/java/com/example/cardinalgate/ui/adapter/CardRecyclerAdapter.java
@@ -1,10 +1,10 @@
 package com.example.cardinalgate.ui.adapter;
 
 import android.content.Context;
-import android.media.Image;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.Button;
+import android.widget.CheckBox;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -13,12 +13,9 @@ import androidx.core.content.res.ResourcesCompat;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.cardinalgate.R;
-import com.example.cardinalgate.core.api.APIClient;
-import com.example.cardinalgate.core.api.APIInterface;
 import com.example.cardinalgate.core.api.model.responses.GetCardsResponse;
-import com.example.cardinalgate.core.callback.CardUnlinkAction;
+import com.example.cardinalgate.core.callback.SetHostCardAction;
 import com.example.cardinalgate.ui.UIHelper;
-import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
 import java.util.ArrayList;
 
@@ -26,15 +23,15 @@ public class CardRecyclerAdapter extends RecyclerView.Adapter<CardRecyclerAdapte
     private final ArrayList<GetCardsResponse.Card> cards;
     private final Context context;
     private final LayoutInflater inflater;
-    private final APIInterface apiClient;
-    private final CardUnlinkAction onCardUnlinked;
+    private final boolean supportsHCEF;
+    private final SetHostCardAction onSetHostCard;
 
-    public CardRecyclerAdapter(Context context, ArrayList<GetCardsResponse.Card> cards, CardUnlinkAction onCardUnlinked) {
+    public CardRecyclerAdapter(Context context, ArrayList<GetCardsResponse.Card> cards, boolean supportsHCEF, SetHostCardAction onSetHostCard) {
         this.context = context;
         this.cards = cards;
         this.inflater = LayoutInflater.from(context);
-        this.apiClient = APIClient.getClient().create(APIInterface.class);
-        this.onCardUnlinked = onCardUnlinked;
+        this.supportsHCEF = supportsHCEF;
+        this.onSetHostCard = onSetHostCard;
     }
 
     @NonNull
@@ -53,9 +50,13 @@ public class CardRecyclerAdapter extends RecyclerView.Adapter<CardRecyclerAdapte
         holder.cardId.setText(card.id);
         holder.addedAt.setText(String.format(context.getString(R.string.card_added_at), card.added));
 
-        holder.unlinkButton.setOnClickListener(v -> {
-            onCardUnlinked.onCardUnlink(card);
+        holder.setHostCard.setOnClickListener(v -> {
+            onSetHostCard.onSetHostCard(card);
         });
+
+        if(!supportsHCEF) {
+            holder.setHostCard.setEnabled(false);
+        }
     }
 
     @Override
@@ -63,11 +64,15 @@ public class CardRecyclerAdapter extends RecyclerView.Adapter<CardRecyclerAdapte
         return cards.size();
     }
 
+    public GetCardsResponse.Card getItem(int position) {
+        return cards.get(position);
+    }
+
     public static class ViewHolder extends RecyclerView.ViewHolder {
         private final ImageView cardIcon;
         private final TextView cardId;
         private final TextView addedAt;
-        private final Button unlinkButton;
+        private final Button setHostCard;
 
         public ViewHolder(@NonNull View itemView) {
             super(itemView);
@@ -75,7 +80,7 @@ public class CardRecyclerAdapter extends RecyclerView.Adapter<CardRecyclerAdapte
             cardIcon = itemView.findViewById(R.id.cardIcon);
             cardId = itemView.findViewById(R.id.cardId);
             addedAt = itemView.findViewById(R.id.addedAt);
-            unlinkButton = itemView.findViewById(R.id.unlinkButton);
+            setHostCard = itemView.findViewById(R.id.setHostCard);
 
             cardIcon.setColorFilter(UIHelper.getColorOnSurface(itemView));
         }